requirements.txt
langchain
langchain-community
openai
groq
tiktoken
pinecone-client
langchain-pinecone
sentence-transformers
numpy
python-dotenv
tqdm